Adding some more resliency
