+++ reflex/utils/frontend_skeleton.py
@@ -299,6 +299,7 @@ def _compile_vite_config(config: Config):
         experimental_hmr=environment.VITE_EXPERIMENTAL_HMR.get(),
         sourcemap=environment.VITE_SOURCEMAP.get(),
         vite_config=config.vite_config,
+        allowed_hosts=config.vite_allowed_hosts,
     )
